+++ src/eca/features/chat.clj
@@ -354,7 +354,7 @@
         rules (f.rules/all config (:workspace-folders db))
         refined-contexts (f.context/raw-contexts->refined contexts db config)
         repo-map* (delay (f.index/repo-map db config {:as-string? true}))
-        instructions (f.prompt/build-instructions refined-contexts rules repo-map* (or behavior (:chat-default-behavior db)) config)
+        instructions (f.prompt/build-instructions refined-contexts rules repo-map* (or behavior (-> config :chat :defaultBehavior)) config)
         chat-ctx {:chat-id chat-id
                   :request-id request-id
                   :contexts contexts